requirements.txt
# Core data science stack
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Financial data
yfinance>=0.2.18
fredapi>=0.5.0
ccxt>=4.0.0

# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration and environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter and development
jupyter>=1.0.0
ipykernel>=6.25.0

# NLP and ML
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0

# Econometrics and statistics
statsmodels>=0.14.0
arch>=6.2.0

# Blockchain data (optional)
web3>=6.0.0

# Data formats
polars>=0.20.0
pyarrow>=12.0.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
